Move connector status check to router config controller
This prevents an issue where the router cannot be reached and the
connector was marked as ready before it was ready since the status
was reset by the router config controller.
